+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Next release
 
 - Add support for the OAuth Applications API. Added the oauthapplication package for API operations and a clerk.OAuthApplication type.
+- Add support for multiple invitation templates with the `TemplateSlug` field in `invitation.Create`.
+- Add support for listing and creating waitlist entries with the `waitlistentry.List` and `waitlistentry.Create` methods.
 
 ## 2.2.0